Doors
I have uploaded the doors for our game so players can progress through to the next level. I have made door assets which can be put on every wall and have additional animations. I have also uploaded the sprites of the doors closed and open, in case the animations are of no use